+++ client/screens/plants.json
@@ -6,7 +6,7 @@
     "imageURL":
       "https://www.petguide.com/wp-content/uploads/2019/03/poison-ivy-dogs-668x444.jpg",
     "description":
-      "A poisonous Asian and Eastern North American flowering plant in the genus Toxicodendron. Causes rashes when touched",
+      "A poisonous Asian and Eastern North American flowering plant in the genus Toxicodendron. Causes rashes when touched.",
     "isPoisonous": true
   }
   // {
@@ -16,7 +16,7 @@
   //   "imageURL":
   //     "http://nativeplantspnw.com/wp-content/uploads/2017/01/Poison-Oak-624x470.png",
   //   "description":
-  //     "A woody vine or shrub in the sumac family, Anacardiaceae. Causes rashes when touched",
+  //     "A woody vine or shrub in the sumac family, Anacardiaceae. Causes rashes when touched.",
   //   "isPoisonous": true
   // },
   // {
@@ -26,7 +26,7 @@
   //   "imageURL":
   //     "https://cdn.britannica.com/50/198250-050-2D0D4360/Poison-sumac-Massachusetts.jpg",
   //   "description":
-  //     "A woody shrub or small tree growing to 9 m (30 ft) tall. Causes a rash when touch",
+  //     "A woody shrub or small tree growing to 9 m (30 ft) tall. Causes a rash when touched.",
   //   "isPoisonous": true
   // },
   // {
@@ -36,7 +36,7 @@
   //   "imageURL":
   //     "https://cdn.mos.cms.futurecdn.net/S7Jyi2NDHX5iCBGrd63Uyh-320-80.jpg",
   //   "description":
-  //     "Commonly known to be the mushroom which Mario Mario's Growth mushroom is based off of, fly agaric is a basidiomycete of the genus Amanita. It is also a muscimol mushroom. It can cause liver toxicity, hallucinations, and death when eaten",
+  //     "Commonly known to be the mushroom which Mario Mario's Growth mushroom is based off of, fly agaric is a basidiomycete of the genus Amanita. It is also a muscimol mushroom. It can cause liver toxicity, hallucinations, and death when eaten.",
   //   "isPoisonous": true
   // }
 ]
